refactor: useAuth 이름 변환 외 2건 (#65)
- useAuth 이름 useLogin으로 변경
- useLogin options prop 추가
- token 함수명 충돌 문제 해결
