+++ src/types/api-response-type/auth-response-type.ts
@@ -1,6 +1,6 @@
 export interface LoginResponse {
-  accessToken: string;
-  refreshToken: string;
+  access_token: string;
+  refresh_token: string;
   user: {
     id: number;
     email: string;
